refactor(tarko): remove meaningless re-exports and restructure web-ui config
