+++ multimodal/tarko/agent-web-ui/src/common/constants/index.ts
@@ -81,14 +81,4 @@ export const CONNECTION_SETTINGS = {
   RECONNECTION_DELAY_MAX: 5000,
 };
 
-/**
- * Agent configuration
- */
-export {
-  getAgentTitle,
-  getWebUIConfig,
-  isContextualSelectorEnabled,
-  getLogoUrl,
-  getWorkspaceNavItems,
-  getGUIAgentConfig,
-} from './shared';
+
